refactor(keybinding-enums): rename enum name to singular nouns and move to a separate file.
