+++ 0_CstSharedResources/Runtime/Scripts/Input Systems/KeybindingAction.cs
@@ -1,0 +1,31 @@
+namespace CST.Shared.Resources
+{
+	public enum KeybindingAction
+	{
+		Attack = 0,
+		ToggleAimMode = 1,
+		Aiming = 2,
+		MoveLeft = 3,
+		MoveRight = 4,
+		MoveUp = 5,
+		MoveDown = 6,
+		Movement = 7,
+		Jump = 13,
+		ContinueDialogue = 8,
+		OpenInventory = 15,
+		Interact = 9,
+		Reload = 10,
+		Mining = 14,
+		BackToMenu = 11,
+		SkipPlayable = 12
+	}
+
+	public enum MouseButtonType
+	{
+		Left,
+		Right,
+		Middle,
+		Forward,
+		Backward
+	}
+}
